# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "먹어도 돼? API",
    "description" : "지병이 있는 사람이 관광지에서 외식할 때, 그 메뉴를 먹어도 되는지 /\n무엇을 빼고 먹어야 하는지 알려주는 서비스의 백엔드 API.\n\n## 인증\n\n자체 ID/PW + JWT. 로그인 응답의 `accessToken` 을 `Authorization: Bearer <토큰>` 으로 보낸다.\n\n- **필수**: `/api/me/**`, `/api/order-cards/**`, `POST /api/reports`\n- **선택**: `/api/restaurants/**` — 토큰이 없으면 목록만, 있으면 **사용자 기준 판정**이 함께 온다.\n  이때 응답의 `personalized` 로 어느 쪽인지 알 수 있다.\n- **불필요**: `/api/auth/**`, `GET /api/reports/options`\n\n## 판정값 (`seal`)\n\n색만으로 전달하지 않는다. 기호와 라벨을 항상 함께 노출할 것.\n\n| verdict | symbol | label |\n|---|---|---|\n| `RED` | ✕ | 알레르기 주의 |\n| `INK` | △ | 조절하면 가능 |\n| `OK` | ○ | 먹어도 돼요! |\n\n메뉴 판정은 **비관적**(하나라도 걸리면 위험), 식당 판정은 **낙관적**(안전한 메뉴가\n하나라도 있으면 갈 만함)이다. 방향이 반대인 것은 의도된 설계다.\n\n`seal` 이 `null` 인 경우가 있다. 아직 **분석되지 않은(미태깅) 메뉴/식당**이라는 뜻이며,\n**안전하다는 뜻이 아니다.** 프론트는 이 경우 판정 배지를 그리지 말고\n\"분석 전\" 으로 표시해야 한다.\n\n## 오류 형식\n\n모든 오류는 `{ \"code\": \"...\", \"message\": \"...\" }` 형태다.\n\n## 주의\n\n이 API 의 판정은 참고용이며 의학적 조언이 아니다. 응답의 `disclaimer` 문구를\n화면에 그대로 노출할 것.\n",
    "contact" : {
      "name" : "먹어도 돼? 백엔드"
    },
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "로컬"
  } ],
...
  "paths" : {
    "/api/me/profile" : {
      "get" : {
        "tags" : [ "2. 내 정보" ],
        "summary" : "내 정보 조회",
        "description" : "`cares`(주의성분) 의 `auto` 가 `true` 면 질환에서 자동으로 따라온 항목이다.\n사용자가 직접 고른 것이 아니므로 화면에서 구분해 보여줄 것.",
        "operationId" : "profile",
        "responses" : {
          "401" : {
            "description" : "토큰이 없거나 만료됨",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "2. 내 정보" ],
        "summary" : "내 정보 수정",
        "description" : "**보낸 항목만 바꾼다.** 필드를 빼거나 `null` 로 보내면 \"그대로 두기\" 다.\n어떤 값을 비우려면 빈 배열 `[]` 또는 빈 문자열을 명시적으로 보낼 것.\n\n질환을 바꾸면 그에 딸린 주의성분(`auto=true`)도 다시 계산된다.",
        "operationId" : "update",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "401" : {
            "description" : "토큰이 없거나 만료됨",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/order-cards" : {
      "post" : {
        "tags" : [ "4. 주문요청카드" ],
        "summary" : "카드 생성",
        "description" : "`requests` 를 **보내지 않으면** 서버가 메뉴 판정과 사용자 상용 문구로 채운다.\n보내면 그것을 그대로 쓴다 — 사용자가 화면에서 직접 고른 결과이기 때문이다.\n\n문구 다듬기는 ai-service 를 거치며, 실패하면 원문 그대로 나간다\n(`CardRequest.polished=false`). 카드 생성 자체는 실패하지 않는다.",
        "operationId" : "create_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "생성됨",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "토큰이 없거나 만료됨",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "식당 또는 메뉴가 없음",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CardResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/auth/signup" : {
      "post" : {
        "tags" : [ "1. 인증" ],
        "summary" : "회원가입",
        "description" : "프론트 5단계에서 모은 값을 한 번에 보낸다. 성공하면 바로 로그인된 상태의\n토큰을 돌려주므로, 가입 후 따로 로그인을 호출할 필요가 없다.\n\n질환·주의성분·알레르기는 **어휘 정본에 있는 값만** 받는다. 철자가 다르면\n판정 시 교집합이 비어 조용히 실패하기 때문이다.",
        "operationId" : "signup",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SignupRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "가입 성공. 토큰을 함께 돌려준다",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "입력값이 규격에 맞지 않음 (code=INVALID_REQUEST)",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409" : {
            "description" : "이미 쓰고 있는 아이디 (code=DUPLICATE_LOGIN_ID)",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login" : {
      "post" : {
        "tags" : [ "1. 인증" ],
        "summary" : "로그인",
        "description" : "아이디가 없는 경우와 비밀번호가 틀린 경우를 **구분해 알리지 않는다.**\n구분해 주면 어떤 아이디가 가입되어 있는지 알아낼 수 있기 때문이다.\n프론트도 두 경우를 같은 문구로 안내할 것.",
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "로그인 성공",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "아이디 또는 비밀번호가 틀림 (code=INVALID_CREDENTIALS)",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/order-cards/{id}" : {
      "get" : {
        "tags" : [ "4. 주문요청카드" ],
        "summary" : "카드 조회",
        "description" : "내가 만든 카드만 볼 수 있다.",
        "operationId" : "get",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "카드 id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          },
          "example" : 1
        } ],
        "responses" : {
          "200" : {
            "description" : "조회됨",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "토큰이 없거나 만료됨",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "없거나 내 카드가 아님",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CardResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ProfileUpdateRequest" : {
        "type" : "object",
        "description" : "프로필 수정 요청. 보낸 항목만 바뀌고, 빠진 항목/null 은 그대로 둔다",
        "properties" : {
          "name" : {
            "type" : "string",
            "maxLength" : 50,
            "minLength" : 0
          },
          "gender" : {
            "type" : "string"
          },
          "birthYear" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 2100,
            "minimum" : 1900
          },
          "bloodType" : {
            "type" : "string"
          },
          "diseases" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          },
          "cares" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          },
          "allergies" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          },
          "chewingDifficulty" : {
            "type" : "boolean"
          },
          "medications" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          },
          "medNote" : {
            "type" : "string"
          },
          "showMedsOnCard" : {
            "type" : "boolean"
          },
          "fontScaleIdx" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 4,
            "minimum" : 0
          }
        }
      },
      "CareView" : {
        "type" : "object",
        "description" : "주의성분 1건",
        "properties" : {
          "value" : {
            "type" : "string",
            "description" : "성분 이름",
            "example" : "나트륨"
          },
          "auto" : {
            "type" : "boolean",
            "description" : "true 면 질환에서 자동으로 따라온 것. 사용자가 고른 것과 구분해 보여줄 것",
            "example" : true
          },
          "note" : {
            "type" : "string",
            "description" : "왜 주의해야 하는지",
            "example" : "고혈압"
          }
        }
      },
      "ProfileResponse" : {
        "type" : "object",
        "description" : "내 정보",
        "properties" : {
          "loginId" : {
            "type" : "string",
            "example" : "meog0112"
          },
          "name" : {
            "type" : "string"
          },
          "gender" : {
            "type" : "string"
          },
          "birthYear" : {
            "type" : "integer",
            "format" : "int32"
          },
          "bloodType" : {
            "type" : "string"
          },
          "diseases" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "cares" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CareView"
            }
          },
          "allergies" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "chewingDifficulty" : {
            "type" : "boolean"
          },
          "medications" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "medNote" : {
            "type" : "string"
          },
          "showMedsOnCard" : {
            "type" : "boolean"
          },
          "fontScaleIdx" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "글자 크기 단계 (0~4)",
            "example" : 0
          },
          "disclaimer" : {
            "type" : "string",
            "description" : "화면에 그대로 노출해야 하는 고지 문구"
          }
        }
      },
      "CreateRequest" : {
        "type" : "object",
        "description" : "제보 등록 요청",
        "properties" : {
          "restaurantId" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "어느 식당에 대한 제보인지",
            "example" : 1
          },
          "ok" : {
            "type" : "boolean",
            "description" : "요청이 받아들여졌는지 — 가게에 대한 평가",
            "example" : true
          },
          "note" : {
            "type" : "string",
            "description" : "자유 메모. 자유 문장은 여기에만 넣을 것",
            "maxLength" : 1000,
            "minLength" : 0
          },
          "requests" : {
            "type" : "array",
            "description" : "그때 실제로 사용한 요청 문구 — 다른 사용자에게 주문 방법으로 보인다",
            "example" : [ "국물은 따로 주세요" ],
            "items" : {
              "type" : "string"
            },
            "maxItems" : 20,
            "minItems" : 0
          },
          "feedback" : {
            "type" : "array",
            "description" : "GET /api/reports/options 의 문구 중에서 고른 것. flag 도출의 근거라\n**문구를 그대로** 보내야 한다. 철자가 다르면 집계에서 빠진다.",
            "example" : [ "덜짜게 해줌" ],
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          }
        },
        "required" : [ "ok", "restaurantId" ]
      },
...
      "CardRequest" : {
        "type" : "object",
        "description" : "카드에 인쇄될 문구 1건",
        "properties" : {
          "phrase" : {
            "type" : "string",
            "description" : "카드에 실제로 찍히는 문구",
            "example" : "국물은 따로 주시면 감사하겠습니다"
          },
          "original" : {
            "type" : "string",
            "description" : "다듬기 전 원문",
            "example" : "국물은 따로 주세요"
          },
          "polished" : {
            "type" : "boolean",
            "description" : "false 면 다듬기에 실패해 원문을 그대로 쓴 것이다. 카드 생성은 실패하지 않는다",
            "example" : true
          }
        }
      },
      "CardResponse" : {
        "type" : "object",
        "description" : "카드 내용. 필드 순서가 카드의 위에서 아래 순서와 같다",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "allergyBanner" : {
            "type" : "string",
            "description" : "맨 위 알레르기 경고 띠. 없으면 null",
            "example" : "새우 알레르기가 있습니다"
          },
          "diseaseLine" : {
            "type" : "string",
            "description" : "지병 줄",
            "example" : "당뇨·고혈압이 있습니다"
          },
          "menuLine" : {
            "type" : "string",
            "description" : "주문할 메뉴 줄",
            "example" : "초당순두부 1인분"
          },
          "requests" : {
            "type" : "array",
            "description" : "요청 문구들",
            "items" : {
              "$ref" : "#/components/schemas/CardRequest"
            }
          },
          "medsLine" : {
            "type" : "string",
            "description" : "복용약 줄. 프로필의 showMedsOnCard 가 false 면 null 이다",
            "example" : "혈압약을 복용 중입니다"
          },
          "footer" : {
            "type" : "array",
            "description" : "카드 아래쪽 안내 문구들",
            "items" : {
              "type" : "string"
            }
          },
          "disclaimer" : {
            "type" : "string",
            "description" : "화면에 그대로 노출해야 하는 고지 문구"
          }
        }
      },
      "SignupRequest" : {
        "type" : "object",
        "description" : "회원가입 요청. 프론트 5단계에서 모은 값을 한 번에 보낸다",
        "properties" : {
          "loginId" : {
            "type" : "string",
            "description" : "로그인 아이디",
            "example" : "meog0112",
            "maxLength" : 50,
            "minLength" : 4
          },
          "password" : {
            "type" : "string",
            "description" : "비밀번호 (6자 이상)",
            "example" : "secret123",
            "maxLength" : 100,
            "minLength" : 6
          },
          "name" : {
            "type" : "string",
            "description" : "이름",
            "example" : "김영수",
            "maxLength" : 50,
            "minLength" : 0
          },
          "gender" : {
            "type" : "string",
            "description" : "성별",
            "example" : "남성"
          },
          "birthYear" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "태어난 해. 카드의 연령대 표기에 쓴다",
            "example" : 1958,
            "maximum" : 2100,
            "minimum" : 1900
          },
          "bloodType" : {
            "type" : "string",
            "description" : "혈액형",
            "example" : "A형"
          },
          "diseases" : {
            "type" : "array",
            "description" : "지병. 어휘 정본에 있는 값만",
            "example" : [ "당뇨", "고혈압" ],
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          },
          "cares" : {
            "type" : "array",
            "description" : "직접 고른 주의성분. 질환에서 자동으로 따라오는 것은 보내지 않아도 된다",
            "example" : [ "나트륨" ],
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          },
          "allergies" : {
            "type" : "array",
            "description" : "알레르기. 어휘 정본에 있는 값만",
            "example" : [ "새우", "고등어" ],
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          },
          "chewingDifficulty" : {
            "type" : "boolean",
            "description" : "씹기 어려움 여부",
            "example" : false
          },
          "medications" : {
            "type" : "array",
            "description" : "복용 중인 약",
            "example" : [ "혈압약" ],
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          },
          "medNote" : {
            "type" : "string",
            "description" : "약에 대한 자유 메모"
          },
          "showMedsOnCard" : {
            "type" : "boolean",
            "description" : "주문요청카드에 복용약을 노출할지",
            "example" : false
          }
        },
        "required" : [ "loginId", "name", "password" ]
      },
      "ApiError" : {
        "type" : "object",
        "description" : "모든 오류 응답의 공통 형식",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "오류 코드. 분기는 이 값으로 할 것",
            "example" : "INVALID_CREDENTIALS"
          },
          "message" : {
            "type" : "string",
            "description" : "사용자에게 보여도 되는 문구",
            "example" : "아이디 또는 비밀번호가 올바르지 않습니다."
          }
        }
      },
      "TokenResponse" : {
        "type" : "object",
        "description" : "발급된 접근 토큰",
        "properties" : {
          "accessToken" : {
            "type" : "string",
            "description" : "이후 요청에 Authorization: Bearer <이 값> 으로 보낸다"
          },
          "expiresIn" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "만료까지 남은 초",
            "example" : 43200
          },
          "tokenType" : {
            "type" : "string",
            "example" : "Bearer"
          }
        }
      },
      "LoginRequest" : {
        "type" : "object",
        "description" : "로그인 요청",
        "properties" : {
          "loginId" : {
            "type" : "string",
            "example" : "meog0112"
          },
          "password" : {
            "type" : "string",
            "example" : "secret123"
          }
        },
        "required" : [ "loginId", "password" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "로그인(`POST /api/auth/login`) 또는 회원가입 응답의 `accessToken` 을 넣는다.\nSwagger UI 오른쪽 위 **Authorize** 에 토큰 값만 붙여 넣으면 된다 (`Bearer ` 접두사는 UI 가 붙인다).",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}